+++ processor/src/test/fixtures/input/com/example/sting/autofragment/MyAutoFragment.java
@@ -1,0 +1,8 @@
+package com.example.sting.autofragment;
+
+import sting.AutoFragment;
+
+@AutoFragment( "MyAutoFragment" )
+public interface MyAutoFragment
+{
+}