+++ processor/src/test/fixtures/input/com/example/sting/autofragment/MyFrameworkModel.java
@@ -1,0 +1,6 @@
+package com.example.sting.autofragment;
+
+@MyFramework
+class MyFrameworkModel
+{
+}