+++ processor/src/test/fixtures/input/com/example/sting/autofragment/MyFrameworkModelImpl.java
@@ -1,0 +1,6 @@
+package com.example.sting.autofragment;
+
+class MyFrameworkModelImpl
+  extends MyFrameworkModel
+{
+}