+++ processor/src/test/fixtures/input/com/example/sting/autofragment/ServiceViaContributeToStingModel.java
@@ -1,0 +1,11 @@
+package com.example.sting.autofragment;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import sting.ContributeTo;
+
+@ContributeTo( "MyAutoFragment" )
+@ArezComponent( dagger = Feature.DISABLE, allowEmpty = true )
+public abstract class ServiceViaContributeToStingModel
+{
+}